+++ buildtools/src/main/java/nl/weeaboo/vn/buildtools/optimizer/OptimizerPreset.java
@@ -1,0 +1,14 @@
+package nl.weeaboo.vn.buildtools.optimizer;
+
+/**
+ * Global quality profile for various optimizers.
+ */
+public enum OptimizerPreset {
+
+    /** Only perform lossless optimizations */
+    LOSSLESS,
+
+    /** Allow lossy optimizations, optimize for quality over file size */
+    QUALITY;
+
+}