+++ buildtools/src/main/java/nl/weeaboo/vn/buildtools/optimizer/image/IImageOperation.java
@@ -2,6 +2,8 @@ package nl.weeaboo.vn.buildtools.optimizer.image;
 
 import java.io.IOException;
 
+import com.google.errorprone.annotations.CheckReturnValue;
+
 import nl.weeaboo.vn.buildtools.optimizer.image.decoder.IImageDecoder;
 import nl.weeaboo.vn.buildtools.optimizer.image.encoder.IImageEncoder;
 
@@ -13,6 +15,11 @@ import nl.weeaboo.vn.buildtools.optimizer.image.encoder.IImageEncoder;
  */
 interface IImageOperation {
 
+    /**
+     * Performs an operation on the image. A new pixmap may be allocated by this method. In that case, the
+     * original pixmap is disposed.
+     */
+    @CheckReturnValue
     ImageWithDef process(ImageWithDef original) throws IOException;
 
 }